% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_scores.R
\name{agg_scores}
\alias{agg_scores}
\alias{print.agg_scores}
\alias{plot.agg_scores}
\title{Aggregate scores}
\usage{
agg_scores(scores, bounds = (0:5)/5, labels = LETTERS[5:1], B = 1e+05)

\method{print}{agg_scores}(x, ...)

\method{plot}{agg_scores}(x, ...)
}
\arguments{
\item{scores}{numeric with values in [0,1]}

\item{bounds}{numeric grade boundaries}

\item{labels}{character grade labels}

\item{B}{numeric number of bootstrap simulations to use}

\item{x}{an object of class "agg_scores"}

\item{...}{additional arguments}
}
\value{
A character string giving the aggregate grade, with
        several attributes containing details from the aggregation.
}
\description{
Given a set of scores on [0,1] calculate an
aggregate score using Bayesian bootstrap methods
}
\examples{
sc <- rbeta(6, 1.5, 3.5)
ag <- agg_scores(sc)
ag
}
