% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_scores.R
\name{cdf}
\alias{cdf}
\alias{cdf.agg_scores}
\alias{qdf}
\alias{qdf.agg_scores}
\alias{plot.cdf}
\alias{plot.qdf}
\title{Distribution function and inverse}
\usage{
cdf(agg, ...)

\method{cdf}{agg_scores}(agg, ...)

qdf(agg, ...)

\method{qdf}{agg_scores}(agg, ...)

\method{plot}{cdf}(x, ..., xlab = "score", ylab = "F(score)", col = "red")

\method{plot}{qdf}(x, ..., xlab = "F(score)", ylab = "score", col = "blue")
}
\arguments{
\item{agg, x}{an object of class "agg_scores"}

\item{...}{additional arguments}

\item{xlab, ylab, col}{graphics parameters}
}
\value{
A function giving either the cdf or its inverse
}
\description{
Cumulative distribution function and its inverse
from the bootstrap simulations conducted as part
of the scores aggregation process.
}
\examples{
sc <- rbeta(6, 1.5, 3.5)
ag <- agg_scores(sc)
pag <- cdf(ag)
qag <- qdf(ag)
pp <- 1:4/5
plot(pag)
segments(qag(pp), 0, qag(pp), pp, lwd = 0.5)
segments(0, pp, qag(pp), pp, lwd = 0.5)
plot(qag)
segments(pp, 0, pp, qag(pp), lwd = 0.5)
segments(0, qag(pp), pp, qag(pp), lwd = 0.5)
}
