% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_score_function.R
\name{cross_points}
\alias{cross_points}
\title{Find cross points}
\usage{
cross_points(guideline, score = c(0.2, 0.8), increasing = FALSE)
}
\arguments{
\item{guideline}{the guideline value, assumed corresponding to 0.5 score}

\item{score}{the score boundaries}

\item{increasing}{logical: is the score transformation increasing?}
}
\value{
A set of grade boundaries in the value scale
}
\description{
Find the points in the value scale corresponding to grade boundaries
in the score scale, assuming the naive value to score transformation
}
\examples{
cross_points(10, c(0.5, (0:5)/5))
cross_points(10, c(0.5, (0:5)/5), TRUE)
}
