% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_scores.R
\name{grades}
\alias{grades}
\alias{grades.agg_scores}
\title{Extract grades}
\usage{
grades(ag, ...)

\method{grades}{agg_scores}(ag, ...)
}
\arguments{
\item{ag}{output of agg_scores()}

\item{...}{other arguments, currently ignored}
}
\value{
A factor giving the grades
}
\description{
Extract the individual grades from an aggregated scores object.
}
\examples{
sc <- rbeta(6, 1.5, 3.5)
ag <- agg_scores(sc)
grades(ag)
}
