% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_score_function.R
\name{make_score_function}
\alias{make_score_function}
\title{Scoring function maker}
\usage{
make_score_function(
  A,
  E,
  guideline,
  s_A = 0.8,
  s_E = 0.2,
  pfn = pnorm,
  qfn = qnorm,
  inverse = FALSE
)
}
\arguments{
\item{A, E}{cutoffs for the A- and E-grads in the value scale}

\item{guideline}{Guideline valuep}

\item{s_A, s_E}{cutoffs for the A- and E-grades in the score scale}

\item{pfn, qfn}{functions to enact the score transformation, and its inverse}

\item{inverse}{logical: should the function be the inverse mapping, score -> value?}
}
\value{
A function to enact the scoring scheme as determined by guideline and cutoffs
}
\description{
Make a scoring function to enact a prescribed scoring scheme.
}
\examples{
Ref <- 10
r <- cross_points(Ref)
fn <- make_score_function(r[1], r[2], Ref, pfn = plogis, qfn = qlogis)
oldPar <- par(yaxs = "i", mar = c(5,4,2,2)+0.1)
curve(fn, xlim = c(Ref/5, 2.5*Ref), ylim = 0:1,
      xlab = "value", ylab = "score")
abline(lty="dashed", h = (1:4)/5, v = r)
curve(1-(log(x/10, base = 2) + 1)/2, add = TRUE, col="blue", xlim = c(5,20))
text(x = rep(2, 5), y = (seq(1, 9, 2))/10, labels = LETTERS[5:1], font = 2)

curve(fn, xlim = c(Ref/4,4*Ref), ylim = 0:1, log="x",
      xlab = "value (log scale)", ylab = "score")
abline(lty="dashed", h = (1:4)/5, v = r)
curve(1-(log(x/10, base = 2) + 1)/2, add = TRUE, col="red", xlim = c(5,20))
text(x = rep(2.5, 5), y = (seq(1, 9, 2))/10, labels = LETTERS[5:1], font = 2)
par(oldPar)
}
